# openapi.yaml
openapi: 3.0.3
info:
  title: full-stack-framework API
  description: Automatically generated OpenAPI spec from Postman collection "full-stack-framework"
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Default local server
paths:
...
  /api/github:
    get:
      summary: Get Github PRs
      operationId: getGithubPRs
      parameters:
        - name: nf
          in: query
          description: Target network function
          required: true
          schema:
            type: string
            enum:
              - free5gc
              - amf
              - ausf
              - bsf
              - chf
              - n3iwf
              - nef
              - nrf
              - nssf
              - pcf
              - smf
              - tngf
              - udm
              - udr
              - upf
              - openapi
              - util
              - nas
              - ngap
              - pfcp
              - aper
            example: amf
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetGithubPRsResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/github/dependency-suggestions:
    post:
      summary: Suggest dependency PRs
      operationId: suggestDependencyPRs
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestDependencySuggestions'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDependencySuggestions'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    NfPr:
      type: object
      required:
        - nfName
        - pr
      properties:
        nfName:
          type: string
          example: amf
        pr:
          type: integer
          example: 100
    LibraryPr:
      type: object
      required:
        - repoName
        - pr
      properties:
        repoName:
          type: string
          enum:
            - openapi
            - util
            - nas
            - ngap
            - pfcp
            - aper
          example: openapi
        pr:
          type: integer
          example: 67
    LibraryPrSuggestion:
      allOf:
        - $ref: '#/components/schemas/LibraryPr'
        - type: object
          properties:
            title:
              type: string
              example: Update NGAP type definitions
            reason:
              type: string
              example: AMF PR #204 mentions openapi and a related library PR is open
            confidence:
              type: string
              example: medium
...
    PR:
      type: object
      required:
        - number
        - title
      properties:
        number:
          type: integer
          example: 212
        title:
          type: string
          example: "gmm/handler: reject EAP-AKA' AuthenticationResponse with no EAPMessage"
    GetGithubPRsResponse:
      type: object
      required:
        - message
        - prs
      properties:
        message:
          type: string
          example: PRs retrieved successfully
        prs:
          type: array
          items:
            $ref: '#/components/schemas/PR'
    RequestDependencySuggestions:
      type: object
      required:
        - nfPrList
      properties:
        nfPrList:
          type: array
          items:
            $ref: '#/components/schemas/NfPr'
    ResponseDependencySuggestions:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Dependency suggestions generated
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/LibraryPrSuggestion'
...